requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
azure-core==1.30.2
azure-eventhub==5.12.1
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-identity==1.17.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
frozenlist==1.4.1
idna==3.7
isodate==0.6.1
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
